Update beta.yml to use Xcode 11.7 instead of 12.3
Required to compile a working binary for arm64e devices.

With Xcode 11 it should work on:
* A12 and newer: on ios 13.x only
* A11 and below: on ios 13.x and 14.x

Reason:
At the time of writing, it seems that Xcode 12 compiles a binary that works on the same versions for A11 and older devices, but works only on iOS 14 for A12 and newer devices (at least in theory). iOS 14 isn't jailbreak-able for A12 devices just yet, so this is a temporary workaround.