Update fetchAppIcon.m to temporarily disable API scraping. [read description]
• Temporarily disable iTunes API scraping to prevent media player from being randomally invoked.
This seems to be a bug in iOS runtime when making NSURL requests from SpringBoard.
This commit will also result into a usage of the system cached app icon, which is much lower resolution and sometimes might show as an "empty" iOS app icon (GitHub, when using an alternate icon is a good option to reproduce this).